converting floats, with tests
This works, but suffers from some rounding errors.

It may be better to simply store the number as a null terminated string on the stack, then use stdlib's conversions. They are probably faster, and more accurate with float. The downside would more stack usage, and if the number happens to be an unusually long string it could crash the stack. There would need to be some safety limit defined as to the maximum length of the input string.