Removed useless 'nocase' version of delimiter splitting
in the unlikely event that alpha characters are used as delimiters, one can simply include both cases in the delimiter list if needed.

Also minor issue fix in strbuf which would return an uninitialised value, but only if you attempted to assign more than INT_MAX characters to a buffer.